moved random helpers to rand.cc
