+++ dynet/nodes-lstm.cc
@@ -1,3 +1,4 @@
+#include "dynet/rand.h"
 #include "dynet/nodes-lstm.h"
 #include "dynet/matrix-multiply.h"
 
@@ -200,20 +201,19 @@ namespace dynet {
     if(weightnoise_std > 0.f){
       Tensor Wx_noisy(Dim({hidden_dim*4, input_dim},1), nullptr, fx.device, fx.mem_pool);
       Wx_noisy.v = static_cast<float*>(scratch_allocator->allocate(Wx_noisy.d.size() * sizeof(float)));
-      std::uniform_int_distribution<> seed_dist(1, 2147483647);
-      Eigen::internal::NormalRandomGenerator<float> normal_rg1(seed_dist(*rndeng));
+      Eigen::internal::NormalRandomGenerator<float> normal_rg1(draw_random_seed());
       Wx_noisy.tvec().device(*dev.edevice) = Wx_noisy.tvec().random(normal_rg1) * weightnoise_std;
       Wx_noisy.tvec().device(*dev.edevice) += Wx->tvec();
 
       Tensor Wh_noisy(Dim({hidden_dim*4, hidden_dim},1), nullptr, fx.device, fx.mem_pool);
       Wh_noisy.v = static_cast<float*>(scratch_allocator->allocate(Wh_noisy.d.size() * sizeof(float)));
-      Eigen::internal::NormalRandomGenerator<float> normal_rg2(seed_dist(*rndeng));
+      Eigen::internal::NormalRandomGenerator<float> normal_rg2(draw_random_seed());
       Wh_noisy.tvec().device(*dev.edevice) = Wh_noisy.tvec().random(normal_rg2) * weightnoise_std;
       Wh_noisy.tvec().device(*dev.edevice) += Wh->tvec();
 
       Tensor b_noisy(Dim({hidden_dim*4, 1},1), nullptr, fx.device, fx.mem_pool);
       b_noisy.v = static_cast<float*>(scratch_allocator->allocate(b_noisy.d.size() * sizeof(float)));
-      Eigen::internal::NormalRandomGenerator<float> normal_rg3(seed_dist(*rndeng));
+      Eigen::internal::NormalRandomGenerator<float> normal_rg3(draw_random_seed());
       b_noisy.tvec().device(*dev.edevice) = b_noisy.tvec().random(normal_rg3) * weightnoise_std;
       b_noisy.tvec().device(*dev.edevice) += b->tvec();
