+++ dynet/tensor.cc
@@ -285,22 +285,6 @@ void TensorTools::randomize_orthonormal(Tensor& val, real scale) {
   } else { throw std::runtime_error("Bad device type"); }
 }
 
-real rand01() {
-  uniform_real_distribution<real> distribution(0, 1);
-  return distribution(*rndeng);
-}
-
-int rand0n(int n) {
-  if (n <= 0) throw std::runtime_error("Integer upper bound is non-positive");
-  int x = rand01() * n;
-  while (n == x) { x = rand01() * n; }
-  return x;
-}
-
-real rand_normal() {
-  normal_distribution<real> distribution(0, 1);
-  return distribution(*rndeng);
-}
 
 #endif
