+++ examples/variational-autoencoder/train_rnnlm-aevb.cc
@@ -9,6 +9,7 @@
 #include "dynet/expr.h"
 #include "dynet/globals.h"
 #include "dynet/io.h"
+#include "dynet/rand.h"
 #include "getpid.h"
 
 #include <iostream>
